fix(m-estimate): cast ROUND args to numeric in opp_value_histogram RPC
The migration failed to run: `ROUND(double precision, integer) does not
exist`. percentile_cont() returns double precision, and Postgres only has
ROUND(numeric, int) — so the bucket_min/bucket_max ROUND(...,2) calls
errored and the function was never created.

Fix: cast the bucket-bound expressions to ::numeric before ROUND.

+++ supabase/migrations/20260726_opp_value_histogram.sql
@@ -43,8 +43,8 @@ LANGUAGE sql STABLE AS $$
   )
   SELECT
     b.n,
-    ROUND(b.lo + (x.bk - 1) * ((b.hi - b.lo) / p_buckets), 2) AS bucket_min,
-    ROUND(b.lo + x.bk * ((b.hi - b.lo) / p_buckets), 2) AS bucket_max,
+    ROUND((b.lo + (x.bk - 1) * ((b.hi - b.lo) / p_buckets))::numeric, 2) AS bucket_min,
+    ROUND((b.lo + x.bk * ((b.hi - b.lo) / p_buckets))::numeric, 2) AS bucket_max,
     x.cnt AS count
   FROM (
     SELECT bucket_idx AS bk, COUNT(*)::bigint AS cnt
